Add get3DPatch to ViewPicker public API and fix bug when getting depth image from OGRE
ViewPicker::get3DPatch is currently a private member function of the ViewPicker class.
It would be very useful to make it public, since this lets us develop plugins that can
 use the orientations of mesh surfaces and point clouds in the 3D scene. This change
moves it from the private API to the public API.

The method in which OGRE packs depth pixel data has changed in the version currently
used by RViz2 so a data ptr indexing fix in the ViewPicker::getPatchDepthImage function
was needed to correctly calculate distances from the scene camera when getting patches
containing more than one pixel.

Signed-off-by: Joe Schornak <joe.schornak@gmail.com>